# Rework BitReader to use byte slices