Currently BitReader requires callers to give ownership of the data. Since it
actually never needs to have ownership, rework this so that BitReader instead
stores a reference to a byte slice. This allows callers not having to do unnecessary
copies of the raw data.

Note that this changes the public API (I haven't updated the changelog accordingly).