@mwlon 2022-02-15
@mcuelenaere Thanks for the PR! I agree this is a good improvement for 0.7. I double-checked that the tests pass and performance is unharmed.

But could you please change the `simple_decompress` function to take a slice as well? I'd also be happy with accepting `AsRef<[u8]>`, but a regular old slice probably causes the least confusion for most users.