@mwlon 2022-02-15
@mcuelenaere congrats! Out of curiosity, what are you using q_compress for?