@mcuelenaere 2022-02-15
> @mcuelenaere congrats! Out of curiosity, what are you using q_compress for?

At Delta we have an autocompleter microservice that runs in a constrained environment (hard limit on CPU time and memory usage). Most of the code is proprietary, but a small part of it is open sourced at https://github.com/mcuelenaere/porigon.

I'm currently prototyping a change where the document IDs (which are fully kept in-memory), that are linked to a key of the FST (array of `uint64`s), are being compressed with q_compress (which has great compression rate!).

I was also contemplating changing the decompression interface of q_compress to be iterator based (to reduce the memory usage by not allocating `Vec<u64>`s), but so far it looks like that is premature/over optimization.