@mwlon 2022-02-15
@mcuelenaere I've also thought about supporting decompression as an iterator or something similar. It requires a somewhat bigger change (probably surfacing `ChunkDecompressor` in the public API), but we could probably make it happen soon if people are interested.