@mcuelenaere 2022-02-15
I justed tested the impact of this PR on query speed and noticed 5x to 10x improvements in request latency!

Examples:

| before | after      |
| ------ | -------- |
| 1.1s     | ~200ms |
| 900ms | 180ms  |
| 192ms | 21ms     |

Is it possible however that you merged this PR **after** you released 0.6.1? Are you planning to release this in a future release? (0.7?)