@mwlon 2022-02-15
@mcuelenaere Yes, 0.6.1 was out before I merged this. I'll release this in 0.7.

Which versions were you comparing? Is the bottleneck of your request a q_compress decompression?